fastapi_vue/services/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
  { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.7"
aerich = "*"
asyncpg = "*"
bcrypt = "*"
fastapi = "*"
passlib = "*"
python-jose = "*"
python-multipart = "*"
tortoise-orm = "*"
uvicorn = "*"

[tool.poetry.dev-dependencies]
black = "*"
pytest = "*"
tox = "*"
pytest-xdist = "*"


[tool.aerich]
tortoise_orm = "src.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
